backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.12
flask-talisman==1.1.0
PyJWT==2.10.1
bcrypt==4.3.0
python-dotenv==1.1.0
gunicorn==23.0.0
pytest==8.3.4
bleach==6.2.0
cryptography==44.0.0
